Create Sale class, start Item.getCurrentPrice()
Created Sale as a new class. Added lists of Sales to Item and
Category, which can be iterated over to find the greatest
applicable discount if multiple Sales apply to a single item.

Started stub for Item.getCurrentPrice() and changed
the name of the price attribute to suggestedRetailPrice. The
plan is for Item.getCurrentPrice() to evaluate all discount
possibilities and return the current price for a given item.