Add functions to Sale, Category, and Item
Added getters and setters to Category class. Added helper functions
to Item class, including the concept of currentPrice, getting it,
and addItemSale, which starts the workflow for adding a new item
sale. A similar function, addCategorySale, was added to Category.
Also added a public constructor to the Sale class as well as an
isSaleValid method, which checks if a Sale would ever be valid
(have the greatest discount) during its life, and otherwise
prevents the Sale from being stored.