Lists of Item and Category Sales working
Currently requires Item.calculateCurrentPrice() to be called whenever
a sale is created. Would prefer to call it automatically on items
and categories only if a sale is currently valid. This would allow
us to use the "currentPriceSet" value and only recalculate the
price if it had not been calculated today.

Also, no code has been added to add these Item and Category Sales to
the database. Will add that in Seller's "add Sale" workflow, and
possibly to Administrator's "add Category sale" workflow, though
it could be added to getters/setters as well.