Add seller workflow to Manager and DisplayUtilities
Created new methods based on Item viewing methods to allow
searching for an item to put on sale. Include logic for creating
a sale.

Logic DOES work BUT does not yet save sales to the database. Can see
impact a current sale *would* have on the price, but it does not yet
persist between runs of the application.